requirements.txt
transformers
faiss-cpu
langchain
cohere
pytorch